chore: rebuild OCCT with the #280 kernel fix; consumer builds warning-free (#281)
Rebuilds OCCT.xcframework (all 3 slices) with a new carried patch, 0002, a
backport of upstream OCCT#1334 ("Data Exchange - initialize STEP writer healing
parameters", merged 2026-07-10) which lands after our V8_0_0_p1 pin (2026-06-16).

That fixes #280 IN THE KERNEL: STEPCAFControl_Controller's ctor replaced the
actor its base class had configured without re-applying SetShapeProcessFlags,
and AutoRecord()s under the same "STEP" name the plain writer resolves by — so
after any XDE STEP read, shape-level writes ran with empty OperationsFlags,
DirectFaces never ran, and faces on indirect surfaces were silently dropped (a
frustum lost its CONICAL_SURFACE and 63% of its volume while still reporting
isValid == true). p1 already defines InitializeMissingParameters() but never
calls it (dead code, and private); the patch is upstream's one-line call.

The v1.9.2 bridge workaround (repairSTEPWriterActor across the 8 shape-level
write paths) is therefore REMOVED — verified by deleting it and confirming the
regression test still passes against the rebuilt binary, i.e. the kernel now
fixes it on its own.

Package.swift's URL binaryTarget is bumped to the v1.10.1 release + new
checksum. Without that, URL-resolving consumers would silently keep the old
kernel while local sibling builds got the fix.

Build warnings 797 -> 0 (#281):
  - 684 -Wdeprecated-declarations -> 0 via OCCT's own OCCT_NO_DEPRECATED opt-out
    on the OCCTBridge target. Deliberately a .define, not .unsafeFlags: SwiftPM
    rejects unsafeFlags in any package consumed as a dependency, which would
    break every consumer. Migrating off the legacy spellings stays open on #281.
  - 23 -Wshorten-64-to-32 -> 0 via explicit static_cast. All 23 were the same
    shape (an NCollection .Size() assigned to int32_t/int). These were NOT latent
    bugs — unlike the quantize() Int32 overflow, a >2^31-element container is not
    reachable here; the casts document intent.
  - Swift hygiene: four dead bindings in SheetMetal.swift, each leftover from a
    superseded approach the surrounding comments already described (stale
    comments removed too; none was an unfinished calculation); Mesh.swift
    var->let. Plus two warnings not in #281 that a cached build had hidden: a
    deprecated String(cString:) in BRepGraph.swift (decoding now stops at the
    bridge's NUL terminator — decoding the whole fixed 128-byte buffer would
    carry NUL padding into the string) and union(with:) in OCCTTest.

Full suite: 4359 tests, 0 failures. swift build: 0 warnings.

+++ Sources/OCCTBridge/src/OCCTBridge_Curve3D.mm
@@ -937,7 +937,7 @@ int32_t OCCTCurve3DEvaluateGrid(OCCTCurve3DRef curve, const double* params, int3
         }
 
         NCollection_Array1<gp_Pnt> results = evaluator.EvaluateGrid(paramArr);
-        int32_t n = results.Size();
+        int32_t n = static_cast<int32_t>(results.Size());
         for (int32_t i = 0; i < n; i++) {
             const gp_Pnt& pt = results.Value(i + 1);
             outXYZ[i*3]   = pt.X();
@@ -962,7 +962,7 @@ int32_t OCCTCurve3DEvaluateGridD1(OCCTCurve3DRef curve, const double* params, in
         }
 
         NCollection_Array1<GeomGridEval::CurveD1> results = evaluator.EvaluateGridD1(paramArr);
-        int32_t n = results.Size();
+        int32_t n = static_cast<int32_t>(results.Size());
         for (int32_t i = 0; i < n; i++) {
             const GeomGridEval::CurveD1& r = results.Value(i + 1);
             outXYZ[i*3]     = r.Point.X();
